+++ src/main/java/racingcar/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 public class Car {
     private final String name;
