+++ src/main/java/racingcar/car/Winners.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import java.util.List;
